Fix endless loop in case of EBADF in select()
If we do not remove a bad fd from the select list we keep getting the
same error from select, and we stop doing any progress on the
communication side. Thus, we forcefully disable all bad fd as soon as
select fails, and we are back in track, progress ensure and everything
seems to work as expected (no leftover events in the event base).

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>